+++ src/main/java/com/lanu/api_trucking_manager/repositories/FreightRepository.java
@@ -3,5 +3,5 @@ package com.lanu.api_trucking_manager.repositories;
 import com.lanu.api_trucking_manager.entities.Freight;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FreightRepository extends JpaRepository<Freight, Integer> {
+public interface FreightRepository extends JpaRepository<Freight, Long> {
 }
