+++ src/main/java/com/lanu/api_trucking_manager/repositories/TaskRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskRepository extends JpaRepository<Task, Integer> {
-    Page<Task> findByFreightFreightId(Integer freightId, Pageable pageable);
+public interface TaskRepository extends JpaRepository<Task, Long> {
+    Page<Task> findByFreightFreightId(Long freightId, Pageable pageable);
 }
